// src/types/usertype.ts
 interface userType {
   role: string;
   fields: string;
   client: boolean;
   _id: string;
   firstName: string;
   lastName: string;
   username: string;
   createdAt: string;
   address: string;
   birthday: string;
   email: string;
   facebook: string;
   github: string;
   info: string;
   instagram: string;
   phoneNumber: string;
   photo: string;
   telegram: string;
   youtube: string;
 }

 export default userType;